+++ src/sys/os/term/grid/mod.rs
@@ -1,0 +1,19 @@
+// devela::sys::os::term::grid
+//
+#![doc = crate::_DOC_SYS_OS_TERM_GRID!()] // public
+#![doc = crate::_doc!(modules: crate::sys::os::term; grid)]
+#![doc = crate::_doc!(flat:"sys")]
+#![doc = crate::_doc!(hr)]
+//
+
+mod color; // TermColor[s|Kind|Mode]
+mod style; // TermStyle[Ext]
+
+crate::structural_mods! { // _mods
+    _mods {
+        pub use super::{
+            color::*,
+            style::*,
+        };
+    }
+}