+++ src/sys/os/term/mod.rs
@@ -1,7 +1,7 @@
 // devela::sys::os::term
 //
 #![doc = crate::_DOC_SYS_OS_TERM!()] // public
-#![doc = crate::_doc!(modules: crate::sys::os; term)]
+#![doc = crate::_doc!(modules: crate::sys::os; term: grid)]
 #![doc = crate::_doc!(flat:"sys")]
 #![doc = crate::_doc!(hr)]
 //
@@ -11,17 +11,18 @@ mod backend; // TermLinux
 mod cap; // TermCaps
 #[cfg(feature = "event")]
 mod event; // TermInputParser
+pub mod grid; // Terminal cell elements, grids, and composition
 mod line; // TermLineMode
 mod metric; // TermSize
 mod render; // TermRenderer
 mod session; // TermSession, TermPollPolicy, TermMode
 
-crate::structural_mods! { // _mods, _crate_internals
+crate::structural_mods! { // _mods, _pub_mods, _crate_internals
     _mods {
         pub use super::{
             ansi::_all::*,
-            cap::_all::*,
             backend::_all::*,
+            cap::_all::*,
             line::_all::*,
             metric::*,
             render::_all::*,
@@ -29,6 +30,11 @@ crate::structural_mods! { // _mods, _crate_internals
         };
         #[cfg(feature = "event")]
         pub use super::event::_all::*;
+    }
+    _pub_mods {
+        pub use super::{
+            grid::_all::*,
+        };
     }
     _crate_internals {
         pub(crate) use super::{